+++ Branch-SDK/Branch-SDK/Core/BNCError.h
@@ -13,10 +13,11 @@ FOUNDATION_EXPORT NSString * const BNCErrorDomain;
 enum {
     BNCInitError = 1000,
     BNCDuplicateResourceError,
-    BNCInvalidReferralCodeError,
+    BNCInvalidPromoCodeError,
     BNCRedeemCreditsError,
     BNCBadRequestError,
     BNCServerProblemError,
+    BNCNilLogError,
     BNCVersionError
 };