+++ Branch-SDK/Branch-SDK/Core/BNCPreferenceHelper.h
@@ -31,8 +31,8 @@
 @property (assign, nonatomic) BOOL isDebug;
 @property (assign, nonatomic) BOOL isConnectedToRemoteDebug;
 @property (assign, nonatomic) NSInteger retryCount;
-@property (assign, nonatomic) NSInteger retryInterval;
-@property (assign, nonatomic) NSInteger timeout;
+@property (assign, nonatomic) NSTimeInterval retryInterval;
+@property (assign, nonatomic) NSTimeInterval timeout;
 
 + (BNCPreferenceHelper *)preferenceHelper;