+++ Branch-TestBed/Branch-SDK-Tests/BNCSystemObserverTests.m
@@ -14,6 +14,10 @@
 
 @interface BNCSystemObserverTests : XCTestCase
 
+@property (strong, nonatomic) id fileManagerMock;
+@property (strong, nonatomic) id docDirAttributesMock;
+@property (strong, nonatomic) id bundleAttributesMock;
+
 @end
 
 @implementation BNCSystemObserverTests
@@ -29,6 +33,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @0);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithNoStoredVersionAndDatesUnder60SecondsApart {
@@ -40,6 +46,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @0);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithNoStoredVersionAndDatesMoreThan60SecondsApart {
@@ -51,6 +59,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @2);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithNoStoredVersionAndNilCreationDate {
@@ -61,6 +71,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @0);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithNoStoredVersionAndNilUpdateDate {
@@ -71,6 +83,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @0);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithEqualStoredAndCurrentVersion {
@@ -84,6 +98,8 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @1);
+
+    [self clearMocks];
 }
 
 - (void)testGetUpdateStateWithNonEqualStoredAndCurrentVersion {
@@ -98,20 +114,28 @@
     NSNumber *updateState = [BNCSystemObserver getUpdateState];
     
     XCTAssertEqualObjects(updateState, @2);
+    
+    [self clearMocks];
 }
 
 
 #pragma mark - Internals
 
 - (void)stubCreationDate:(NSDate *)creationDate modificationDate:(NSDate *)modificationDate {
-    id fileManagerMock = OCMClassMock([NSFileManager class]);
-    id docDirAttributesMock = OCMClassMock([NSDictionary class]);
-    id bundleAttributesMock = OCMClassMock([NSDictionary class]);
-    [[[fileManagerMock stub] andReturn:fileManagerMock] defaultManager];
-    [[[fileManagerMock expect] andReturn:docDirAttributesMock] attributesOfItemAtPath:[OCMArg any] error:(NSError __autoreleasing **)[OCMArg anyPointer]];
-    [[[fileManagerMock expect] andReturn:bundleAttributesMock] attributesOfItemAtPath:[OCMArg any] error:(NSError __autoreleasing **)[OCMArg anyPointer]];
-    [[[docDirAttributesMock stub] andReturn:creationDate] fileCreationDate];
-    [[[bundleAttributesMock stub] andReturn:modificationDate] fileModificationDate];
+    self.fileManagerMock = OCMClassMock([NSFileManager class]);
+    self.docDirAttributesMock = OCMClassMock([NSDictionary class]);
+    self.bundleAttributesMock = OCMClassMock([NSDictionary class]);
+    [[[self.fileManagerMock stub] andReturn:self.fileManagerMock] defaultManager];
+    [[[self.fileManagerMock expect] andReturn:self.docDirAttributesMock] attributesOfItemAtPath:[OCMArg any] error:(NSError __autoreleasing **)[OCMArg anyPointer]];
+    [[[self.fileManagerMock expect] andReturn:self.bundleAttributesMock] attributesOfItemAtPath:[OCMArg any] error:(NSError __autoreleasing **)[OCMArg anyPointer]];
+    [[[self.docDirAttributesMock stub] andReturn:creationDate] fileCreationDate];
+    [[[self.bundleAttributesMock stub] andReturn:modificationDate] fileModificationDate];
+}
+
+- (void)clearMocks {
+    [self.fileManagerMock stopMocking];
+    [self.docDirAttributesMock stopMocking];
+    [self.bundleAttributesMock stopMocking];
 }
 
 - (void)stubNilValuesForStoredAndCurrentVersions {